Tx size is back to uint32

+++ core/store/ledgerstore/block_store.go
@@ -133,12 +133,12 @@ func (this *BlockStore) loadHeaderWithTx(blockHash common.Uint256) (*types.Heade
 	if err != nil {
 		return nil, nil, err
 	}
-	txSize, eof := source.NextUint64()
+	txSize, eof := source.NextUint32()
 	if eof {
 		return nil, nil, io.ErrUnexpectedEOF
 	}
 	txHashes := make([]common.Uint256, 0, int(txSize))
-	for i := uint64(0); i < txSize; i++ {
+	for i := uint32(0); i < txSize; i++ {
 		txHash, eof := source.NextHash()
 		if eof {
 			return nil, nil, io.ErrUnexpectedEOF
@@ -154,7 +154,7 @@ func (this *BlockStore) SaveHeader(block *types.Block) error {
 	key := this.getHeaderKey(blockHash)
 	sink := common.NewZeroCopySink(nil)
 	block.Header.Serialization(sink)
-	sink.WriteUint64(uint64(len(block.Transactions)))
+	sink.WriteUint32(uint32(len(block.Transactions)))
 	for _, tx := range block.Transactions {
 		txHash := tx.Hash()
 		sink.WriteHash(txHash)
